+++ Rubjerg.Graphviz.Test/CGraphEdgeCases.cs
@@ -1,5 +1,4 @@
 using System.Linq;
-using System.Runtime.InteropServices;
 using NUnit.Framework;
 
 namespace Rubjerg.Graphviz.Test
@@ -7,17 +6,6 @@ namespace Rubjerg.Graphviz.Test
     [TestFixture()]
     public class CGraphEdgeCases
     {
-        [DllImport("GraphvizWrapper.dll", SetLastError = true, CharSet = CharSet.Ansi, CallingConvention = CallingConvention.Cdecl)]
-        protected static extern void rj_debug();
-
-        [Test()]
-        [Ignore("For debugging")]
-        public void debug()
-        {
-            NativeMethods.CreateConsole();
-            rj_debug();
-        }
-
         [Test()]
         public void TestAttributeIntroduction()
         {
