+++ Rubjerg.Graphviz.Test/GraphvizBasicOperations.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -78,35 +78,6 @@ namespace Rubjerg.Graphviz.Test
             // Because Graphviz uses a lower-left originated coordinate system, we need to flip the y coordinates
             Utils.AssertOrder(rects, r => (r.Left, -r.Top));
             Assert.That(rects.Count, Is.EqualTo(9));
-        }
-
-        /// <summary>
-        /// This test used to fail: https://gitlab.com/graphviz/graphviz/-/issues/1894
-        /// It still fails on github hosted VMs: https://gitlab.com/graphviz/graphviz/-/issues/1905
-        /// </summary>
-        [Test()]
-        [TestCase("Times-Roman", 7, 0.01)]
-        [TestCase("Times-Roman", 7, 0.5)]
-        [Category("Flaky")]
-        public void TestRecordShapeAlignment(string fontname, double fontsize, double margin)
-        {
-            RootGraph root = CreateUniqueTestGraph();
-            // Margin between label and node boundary in inches
-            Node.IntroduceAttribute(root, "margin", margin.ToString(CultureInfo.InvariantCulture));
-            Node.IntroduceAttribute(root, "fontsize", fontsize.ToString(CultureInfo.InvariantCulture));
-            Node.IntroduceAttribute(root, "fontname", fontname);
-
-            Node nodeA = root.GetOrAddNode("A");
-
-            nodeA.SafeSetAttribute("shape", "record", "");
-            nodeA.SafeSetAttribute("label", "{20 VH|{1|2}}", "");
-
-            //TestContext.Write(root.ToDotString());
-            root.ComputeLayout();
-            //TestContext.Write(root.ToDotString());
-
-            var rects = nodeA.GetRecordRectangles().ToList();
-            Assert.That(rects[0].Right, Is.EqualTo(rects[2].Right));
         }
 
         [Test()]
